Crate rattler_redaction

Constants

DEFAULT_REDACTION_STR

A default string to use for redaction.

Traits

Redact

A trait to redact known secrets from a type.

Functions

redact_known_secrets_from_url

Anaconda channels are not always publicly available. This function checks if a URL contains a secret by identifying whether it contains certain patterns. If it does, the function returns a modified URL where any secret has been masked.